// src/app/admin-view/models/addoffer.ts
export class AddOffer {
    OfferID: number;
    StateId: string;
    StateName: string;
    DistId: string;
    DistName: string;
    LocalBodyId: string;
    LocalBodyName: string;
    PartyTypeID: string;
    PartyID: string;
    OfferNo: string;
    OfferHeading: string;
    OfferCode: string;
    OfferStartDate: string;
    OfferEndDate: string;
    MobileNo: string;
    TokenId: string;
}